+++ wayland-scanner/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Generated protocol object types now implement `Borrow<ObjectId>`
+
 ## 0.30.0 -- 27/12/2022
 
 ## 0.30.0-beta.10